Improve MSVC toolchain compatibility

+++ src/libstfio/hdf5/hdf5lib.cpp
@@ -59,7 +59,7 @@ bool stfio::exportHDF5File(const std::string& fName, const Recording& WData, Pro
 
     /* Define an array of root tables */
     rt p_data;
-    p_data.channels = WData.size();
+    p_data.channels = static_cast<int>(WData.size());
     struct tm t = WData.GetDateTime();
     std::size_t date_length = snprintf(p_data.date, DATELEN, "%04i-%02i-%02i", t.tm_year+1900, t.tm_mon+1, t.tm_mday);
     std::size_t time_length = snprintf(p_data.time, TIMELEN, "%02i:%02i:%02i", t.tm_hour, t.tm_min, t.tm_sec);
